TASK005 1. Исправлено некорректное отображение AvatarView на странице со списком друзей. 2. Добавлен CustomLikeControl для фото друзей.
